include test programs in make for the retro, rename patterns to nabu-games-patterns
