+++ include/nabu-games-patterns.h
